+++ cloudfeaster/jsonschemas.py
@@ -16,4 +16,4 @@ def _load_jsonschema(schema_name):
 
 spider_metadata = _load_jsonschema('spider_metadata')
 
-spider_output = _load_jsonschema('spider_output')
+crawl_result = _load_jsonschema('crawl_result')