+++ cloudfeaster/tests/privacy_tests.py
@@ -9,8 +9,8 @@ from .. import privacy
 class TestHashSpiderArg(unittest.TestCase):
 
     def test_ok_no_data(self):
-        spider_arg = 'something sensitive'
-        hashed_spider_arg = privacy.hash_spider_arg(spider_arg)
-        self.assertNotEqual(spider_arg, hashed_spider_arg)
+        crawl_arg = 'something sensitive'
+        hashed_crawl_arg = privacy.hash_crawl_arg(crawl_arg)
+        self.assertNotEqual(crawl_arg, hashed_crawl_arg)
         reg_ex = re.compile(r'^[^\s]+:[^\s]+$')
-        self.assertTrue(reg_ex.match(hashed_spider_arg))
+        self.assertTrue(reg_ex.match(hashed_crawl_arg))